Switch to Int64 for ASN1 integer fields, as webOrderLineItemID requires this

+++ AppReceiptValidator/AppReceiptValidator/OpenSSL/ASN1Helpers.swift
@@ -155,7 +155,7 @@ extension ASN1Object {
 
 extension ASN1Object {
 
-    var intValue: Int? {
+    var intValue: Int64? {
         guard self.type == V_ASN1_INTEGER else { return nil }
 
         var pointer = self.valuePointer
@@ -163,11 +163,11 @@ extension ASN1Object {
         defer {
             ASN1_INTEGER_free(integer)
         }
-        let result = ASN1_INTEGER_get(integer)
+        let result = Int64(ASN1_INTEGER_get(integer))
         return result
     }
 
-    func intValue(byAdvancingPointer pointer: inout UnsafePointer<UInt8>?, length: Int? = nil) -> Int? {
+    func intValue(byAdvancingPointer pointer: inout UnsafePointer<UInt8>?, length: Int? = nil) -> Int64? {
         let length = length ?? self.length
         pointer = pointer?.advanced(by: length)
         guard let intValue = self.intValue else { return nil }
